Validate annex Output File

Validate XLSX file created by annex_write_stats(). Checks if all required sheets/columns are available and that all user-modified META information has been entered correctly.

Usage

annex_validate(file, user, quiet = FALSE, ...)

Arguments

file

name of the file to be validated (XLSX file).

user

positive integer, the user identifier given by the project team.

quiet

logical, defaults to FALSE. If TRUE, the output will be limited.

...

currently unused.

Value

Some checks will cause an error and stop execution. Others will cause a message with some information on what has to be fixed to make the document valid. If the function does not stop due to an error it will return TRUE if the file has been validated as proper, or FALSE if issues have been found.

Author

Reto Stauffer