Variable definition information

The template not only contains the definition of the allowed variables, it also states whether or not additional information is required (or optional), an upper and lower bound to be considered 'valid' plus (is specified) a series of allowed units. Used to prepare the data and convert to annex standard units, quality checks, as well as validation.

Usage

annex_variable_definition(as_list = FALSE)

Arguments

as_list

logical. If FALSE (default) a data.frame will be returned, if TRUE a list (see Details).

Value

Returns either a data.frame or list of lists

which contains the allowed (defined) variables.

Details

If as_list = TRUE a list of lists is returned, else a data.frame.

List: The name of the list corresponds to the name of the variable, whereas each entry contains a list with a logical flag if additional information in the META sheet is required as well as a numeric lower and upper bound which defines in which range an observation is considered to be valid. Can be NA if not specified (both or one of them). allowed_units contains NA (unspecified) or a character wich one or multiple comma separated units specifications.

If as_list = FALSE (default) the same information is returned as a data.frame containing the same information.

See also

Author

Reto Stauffer